requirements.txt
langchain
langchain-openai
pydantic
python-dotenv
networkx
matplotlib
tiktoken
tqdm
pandas
spacy
plotly